[axl.gui] bugfix: register format should be uintptr_t, not uint_t

+++ src/axl_gui/axl_gui_HyperText.cpp
@@ -109,7 +109,7 @@ CHyperText::AppendHyperText (
 		}
 		else if (pTag[1] == '\b') // backspace
 		{
-			size_t BackLength = pTag[2] != '*' ? strtoul (pTag + 2, NULL, 10) : -1;
+			size_t BackLength = pTag[2] != '*' ? (size_t) strtoul (pTag + 2, NULL, 10) : -1;
 			Backspace (BackLength ? BackLength : 1);
 		}
 		else
